Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    int a,n;
-    scanf("%d %d",&a,&n);
+    float a,n;
+    scanf("%f %f",&a,&n);
     if(a/n==n){
         printf("Yes");}
     else printf("No");
